move colourMapMenu in a separate file
